# tests/characterization/openapi_reference.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OverviewOut": {
        "properties": {
          "approved_shops": {
            "title": "Approved Shops",
            "type": "integer"
          },
          "pending_shops": {
            "title": "Pending Shops",
            "type": "integer"
          },
          "rejected_shops": {
            "title": "Rejected Shops",
            "type": "integer"
          },
          "total_invoices": {
            "title": "Total Invoices",
            "type": "integer"
          },
          "total_revenue": {
            "title": "Total Revenue",
            "type": "number"
          },
          "total_shops": {
            "title": "Total Shops",
            "type": "integer"
          }
        },
        "required": [
          "total_shops",
          "pending_shops",
          "approved_shops",
          "rejected_shops",
          "total_invoices",
          "total_revenue"
        ],
        "title": "OverviewOut",
        "type": "object"
      },
...
      "RejectRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "RejectRequest",
        "type": "object"
      },
...
      "ShopAdminOut": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Email"
          },
          "owner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "created_at"
        ],
        "title": "ShopAdminOut",
        "type": "object"
      },
      "ShopListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ShopAdminOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "ShopListOut",
        "type": "object"
      },
...
      "ShopStatsOut": {
        "properties": {
          "clients_count": {
            "title": "Clients Count",
            "type": "integer"
          },
          "invoices_count": {
            "title": "Invoices Count",
            "type": "integer"
          },
          "products_count": {
            "title": "Products Count",
            "type": "integer"
          },
          "shop_id": {
            "title": "Shop Id",
            "type": "integer"
          },
          "shop_name": {
            "title": "Shop Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_revenue": {
            "title": "Total Revenue",
            "type": "number"
          },
          "users_count": {
            "title": "Users Count",
            "type": "integer"
          }
        },
        "required": [
          "shop_id",
          "shop_name",
          "status",
          "products_count",
          "clients_count",
          "invoices_count",
          "total_revenue",
          "users_count"
        ],
        "title": "ShopStatsOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Sunu Boutik API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/overview": {
      "get": {
        "operationId": "overview_admin_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/shops": {
      "get": {
        "operationId": "list_shops_admin_shops_get",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Shops",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/shops/{shop_id}/approve": {
      "post": {
        "operationId": "approve_shop_admin_shops__shop_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "title": "Shop Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Approve Shop",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/shops/{shop_id}/reject": {
      "post": {
        "operationId": "reject_shop_admin_shops__shop_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "title": "Shop Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reject Shop",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/shops/{shop_id}/stats": {
      "get": {
        "operationId": "shop_stats_admin_shops__shop_id__stats_get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "title": "Shop Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopStatsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Shop Stats",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}